refactor: split units into individual files for parallel development
Split regular.ts and elite.ts into individual unit files organized
by unit type (regular/elite) to prevent merge conflicts during
parallel development:

- regular/: 15 individual files for level 1-2 units
- elite/: 12 individual files for level 3-4 units
- Each subdirectory has an index.ts aggregator
- Main index.ts updated to import from subdirectories
- Individual unit definitions now directly exportable
