+++ packages/shared/src/units/regular/northernMonks.ts
@@ -1,0 +1,29 @@
+/**
+ * Northern Monks unit definition
+ */
+
+import { ELEMENT_ICE } from "../../elements.js";
+import type { UnitDefinition } from "../types.js";
+import {
+  UNIT_TYPE_REGULAR,
+  RECRUIT_SITE_MONASTERY,
+  UNIT_ABILITY_ATTACK,
+  UNIT_ABILITY_BLOCK,
+} from "../constants.js";
+import { UNIT_NORTHERN_MONKS } from "../ids.js";
+
+export const NORTHERN_MONKS: UnitDefinition = {
+  id: UNIT_NORTHERN_MONKS,
+  name: "Northern Monks",
+  type: UNIT_TYPE_REGULAR,
+  level: 2,
+  influence: 7,
+  armor: 4,
+  resistances: [],
+  recruitSites: [RECRUIT_SITE_MONASTERY],
+  abilities: [
+    { type: UNIT_ABILITY_ATTACK, value: 4, element: ELEMENT_ICE },
+    { type: UNIT_ABILITY_BLOCK, value: 3, element: ELEMENT_ICE },
+  ],
+  copies: 1,
+};
